tests/requirements.txt
pytest-cov==2.12.1
pytest-subprocess==1.1.2